pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aech-cli-documents"
version = "0.2.0"
description = "Universal Document Corpus CLI for Agent Aech - VLM-powered document intelligence"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer",
    "pdf2image",
    "Pillow",
    "anthropic",
    "sentence-transformers",
    "numpy",
    "rich",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[project.scripts]
aech-cli-documents = "aech_cli_documents.main:run"

[tool.setuptools.package-data]
aech_cli_documents = ["manifest.json"]

[tool.setuptools.packages.find]
where = ["."]
